Fix UpdateOutput of logger by updating the correct outputs only.
This change fixes the updating of outputs of a logger. Each outout
has a set of classes for which it is supposed to be applied on a line
data. In this change an intersection of classes is taken between an
output's classes and given classes to be updated. If intersection is
same as output's set of classes then whole output is updated otherwise
output is split into two, one for intersection and other for the rest.
Intersection one is updated with new URI but old one still keeps old URI.