Use correct forward slashes when dealing with callers path
The filename returned from runtime.Caller is forward slash
delimited which has had an open issue for years:
https://github.com/golang/go/issues/3335

This changes makes it such that the file:line is reported
properly instead of the full-path:line which is the current
behavior on Windows. When dealing with small log lines,
that long path can represent a sizable percentage of the
total size of the log file.